# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service library with HTTP API endpoints",
        "title": "Miko Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/cookiecloud/identity": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the CookieCloud server identity using the provided key and password. This allows the service to fetch cookies from your CookieCloud server for authentication with music platforms.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cookiecloud"
                ],
                "summary": "Update CookieCloud identity",
                "parameters": [
                    {
                        "description": "CookieCloud identity (key and password)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CookiecloudIdentityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Identity updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.CookiecloudIdentityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid JSON or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to update identity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cookiecloud/pull": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Forces the service to pull the latest cookies from your CookieCloud server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cookiecloud"
                ],
                "summary": "Force pull cookies",
                "responses": {
                    "200": {
                        "description": "Cookies pulled successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to pull cookies",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cookiecloud/server": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the currently configured CookieCloud server URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cookiecloud"
                ],
                "summary": "Get CookieCloud server URL",
                "responses": {
                    "200": {
                        "description": "CookieCloud server URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "url": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/download": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Download single or multiple music tracks by providing song IDs, album URLs, playlist URLs, or song URLs. Supports batch downloads with various quality levels and conflict resolution policies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Download music tracks",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": "\"2161154646\"",
                        "description": "Resource URIs to download (can be song ID, album URL, playlist URL, song URL). Multiple URIs supported for batch downloads.",
                        "name": "uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "standard",
                            "higher",
                            "exhigh",
                            "lossless",
                            "hires",
                            "128",
                            "192",
                            "320",
                            "HQ",
                            "SQ",
                            "HR"
                        ],
                        "type": "string",
                        "default": "lossless",
                        "example": "\"hires\"",
                        "description": "Audio quality level",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"./downloads\"",
                        "description": "Output directory path for downloaded files",
                        "name": "output",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 60000,
                        "example": 60000,
                        "description": "Timeout in milliseconds for the download operation, 0 means no timeout",
                        "name": "timeout",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "skip",
                            "overwrite",
                            "rename",
                            "update_tags"
                        ],
                        "type": "string",
                        "default": "skip",
                        "example": "\"skip\"",
                        "description": "How to handle existing files",
                        "name": "conflict_policy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"netease\"",
                        "description": "Music platform to use for downloading",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful batch download response with individual song results and error details",
                        "schema": {
                            "$ref": "#/definitions/models.DownloadSummary"
                        }
                    },
                    "400": {
                        "description": "Bad request - missing or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during download processing",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates a user with username and password and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/platform/{platform}/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's information from the specified music platform",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user information",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"netease\"",
                        "description": "Music platform name",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/types.User"
                        }
                    },
                    "400": {
                        "description": "Bad request - provider creation failed",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - authentication failed or invalid platform",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_stkevintan_miko_pkg_types.Album": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "picUrl": {
                    "type": "string"
                }
            }
        },
        "github_com_stkevintan_miko_pkg_types.Artist": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CookiecloudIdentityRequest": {
            "description": "platform auth request",
            "type": "object",
            "required": [
                "key",
                "password"
            ],
            "properties": {
                "key": {
                    "type": "string",
                    "example": "your-cookiecloud-key"
                },
                "password": {
                    "type": "string",
                    "example": "your-cookiecloud-password"
                }
            }
        },
        "models.CookiecloudIdentityResponse": {
            "description": "platform auth response",
            "type": "object",
            "required": [
                "key",
                "url"
            ],
            "properties": {
                "key": {
                    "type": "string",
                    "example": "your-cookiecloud-key"
                },
                "url": {
                    "type": "string",
                    "example": "https://cookiecloud.example.com"
                }
            }
        },
        "models.DownloadSummary": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.DownloadResult"
                    }
                },
                "summary": {
                    "type": "string",
                    "example": "Downloaded 8 out of 10 songs."
                }
            }
        },
        "models.ErrorResponse": {
            "description": "Error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "types.DownloadResult": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.DownloadedMusic"
                },
                "error": {}
            }
        },
        "types.DownloadedMusic": {
            "type": "object",
            "properties": {
                "album": {
                    "$ref": "#/definitions/github_com_stkevintan_miko_pkg_types.Album"
                },
                "artist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stkevintan_miko_pkg_types.Artist"
                    }
                },
                "filePath": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lyrics": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quality": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                },
                "trackNumber": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "types.User": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}